requirements.txt
# Stacking Sats Challenge - Python Dependencies
# ===============================================

# Core data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# HTTP requests for data downloading
requests>=2.28.0

# Parquet file support
pyarrow>=10.0.0

# Environment variable management
python-dotenv>=0.19.0

# Optional plotting dependencies (lazy loaded)
matplotlib>=3.5.0

# Development and testing (optional)
pytest>=7.0.0
ruff>=0.1.0
autopep8>=2.0.0 